alunos api

+++ src/main/java/com/allan/spr/repositories/CidadeRepository.java
@@ -11,11 +11,9 @@ import org.springframework.transaction.annotation.Transactional;
 import com.allan.spr.domain.Cidade;
 
 @Repository
-public interface CidadeRepository extends JpaRepository <Cidade, Integer> {
-	//Repository são classes DAO.
-	
+public interface CidadeRepository extends JpaRepository <Cidade, Long> {
 	@Transactional(readOnly=true)
 	@Query("SELECT obj FROM Cidade obj WHERE obj.estado.id = :estadoId ORDER BY obj.nome")
-	public List<Cidade> findCidades(@Param("estadoId") Integer estado_id);
+	public List<Cidade> findCidades(@Param("estadoId") Long estado_id);
 
 }
