+++ src/main/java/com/allan/spr/security/UserSS.java
@@ -14,7 +14,7 @@ public class UserSS implements UserDetails {
 	private static final long serialVersionUID = 1L;
 	
 	
-	private Integer id;
+	private Long id;
 	private String email;
 	private String senha;
 	private Collection<? extends GrantedAuthority> authorities;
@@ -25,7 +25,7 @@ public class UserSS implements UserDetails {
 	
 	
 	
-	public UserSS(Integer id, String email, String senha,Set<Perfil> perfis) {
+	public UserSS(Long id, String email, String senha,Set<Perfil> perfis) {
 		super();
 		this.id = id;
 		this.email = email;
@@ -35,7 +35,7 @@ public class UserSS implements UserDetails {
 
 
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 		
 	}
